float to compatible onnx runtime calculation

+++ TreeOfThought/backend/onnx-computer-vision/Services/IInsightFaceSkiaService.cs
@@ -9,6 +9,6 @@ public interface IInsightFaceSkiaService : IDisposable
 {
     List<FaceInfo> DetectFace(SKBitmap bitmap, float detThresh = 0.5f);
     float[] VectorFace(SKBitmap bitmap, FaceInfo face);
-    double CompareVector(float[] v1, float[] v2);
-    (double Similarity, FaceInfo? f1, FaceInfo? f2, float[]? v1, float[]? v2) CompareFaces(SKBitmap img1, SKBitmap img2, float detThresh = 0.5f);
+    float CompareVector(float[] v1, float[] v2);
+    (float Similarity, FaceInfo? f1, FaceInfo? f2, float[]? v1, float[]? v2) CompareFaces(SKBitmap img1, SKBitmap img2, float detThresh = 0.5f);
 }
